requirements.txt
django==1.11.15
django-admin-rangefilter==0.3.7
django-admin-view-permission==1.7
django-countries==5.3.1
django-flat-responsive==2.0
django-import-export==1.0.1
django-money==0.14.2
django-mptt==0.9.0
django-phonenumber-field==2.0.0
django-search-admin-autocomplete==0.1.2
ipython==5.8.0
Pillow==5.2.0
psycopg2==2.7.5
